# dictionary-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dictionary API",
        "description": "Documentação da API de Dicionário",
        "contact": {
            "email": "support@coodesh.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/api/auth/signup": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrar um novo usuário",
                "description": "Permite o registro de um novo usuário na aplicação.",
                "operationId": "6cf11ae1f0353336f0782d896b80cb94",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "User 1"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "example@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "test1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário registrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "f3a10cec013ab2c1380acef"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "User 1"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "Bearer JWT.Token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/signin": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Autenticar um usuário",
                "description": "Permite que um usuário existente faça login.",
                "operationId": "a3c2687dfbd4ce13da7c1d2be8b332ae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "example@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "test1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário autenticado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "f3a10cec013ab2c1380acef"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "User 1"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "Bearer JWT.Token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Credenciais inválidas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/me/favorites": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Listar palavras favoritas",
                "description": "Retorna uma lista de palavras que o usuário marcou como favoritas com paginação.",
                "operationId": "7f10cbad1e2f064ee347467f5205146b",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Número de resultados por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página para paginação",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de palavras favoritas retornada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "word": {
                                                        "type": "string",
                                                        "example": "example"
                                                    },
                                                    "added": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-12-23T12:34:56+00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "totalDocs": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "totalPages": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "hasNext": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "hasPrev": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuário não autenticado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário não autenticado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}